refactor(handlers): extract write.ts handler from intent.ts (Stage B)
handleSAPWrite (the 1.8K-line handler) + its write-only helpers
(isDeleteDependencyError, normalize*Override, TABL_DT hint). Imports
resolveVersionAndDraftInfo from read.ts (acyclic). intent.ts is now ~700
lines (dispatch + error tree + barrel). Last handler extraction.
3672 tests, A1 zero-diff.
